refactor: extract Interfaces into common
