--- packages/puppeteer-core/src/cdp/ConnectOptions.ts
@@ -1,34 +1,0 @@
-/*
- * Copyright 2023 Google Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type {ConnectionTransport} from '../common/ConnectionTransport.js';
-
-import type {BrowserConnectOptions} from './BrowserConnector.js';
-
-/**
- * @public
- */
-export interface ConnectOptions extends BrowserConnectOptions {
-  browserWSEndpoint?: string;
-  browserURL?: string;
-  transport?: ConnectionTransport;
-  /**
-   * Headers to use for the web socket connection.
-   * @remarks
-   * Only works in the Node.js environment.
-   */
-  headers?: Record<string, string>;
-}
